This is the powerpoint presentation of the process.
One of the things to note about this project is that the purpose of these motor fixtures is to help maintain the vibration from cold steel, which helps reduces the chance of carpal tunnel. Riding motocross and ATV bikes is an extreme sport, and the manufacturing process must be efficient and effective with safety concerns.